src/main/java/com/nhnacademy/shoppingmallservice/service/impl/CookieServiceImpl.java
package com.nhnacademy.shoppingmallservice.service.impl;

import com.nhnacademy.shoppingmallservice.service.CookieService;
import jakarta.servlet.http.Cookie;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.stereotype.Service;

@Service
public class CookieServiceImpl implements CookieService {

    @Override
    public void saveOnCookie(HttpServletResponse response, String key, String value, int expiry) {
        Cookie cookie = new Cookie(key, value);

        cookie.setPath("/");
        cookie.setHttpOnly(true);
        cookie.setMaxAge(expiry);

        response.addCookie(cookie);
    }
}
